changed totalconnections to avg connection along with some html
its not perfect but it works pretty well. Also trying to make the html look presentable

+++ sorter_site/Webapp_root/script.js
@@ -29,7 +29,7 @@ $("#Create-canvas-btn").click(function () {
     document.getElementById("currentSeed").textContent = (randomSeed).toString();
     Math.seedrandom(randomSeed);
     var totalNodes = parseInt(document.getElementById("totalNodes").value);
-    var totalConnections = parseInt(document.getElementById("totalConnections").value);
+    var totalConnections = parseFloat(document.getElementById("totalConnections").value);
     var isItSimulated = document.querySelector('#isItSimulated').checked;
     current_screen.startup(totalNodes, totalConnections, isItSimulated,true,false);
